+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.17 (09-05-2026)
 - Implemented a new map viewer tool for exploring maps in 3D directly in wow.export.
 - Implemented a new "ADT LoD" option for terrain exporting which uses the in-game LoD textures.
+- Implemented a new "Export Minimap Tiles" option to the map exporter.
 
 ## 0.2.16 (29-04-2026)
 - Fixed a regression causing terrain UV maps to be broken.